fix: break lutaml/model ↔ format require cycles via internal format files
ruby -w reported 'loading in progress, circular require considered
harmful' for key_value.rb, json.rb and xml.rb against model.rb:
model.rb's bottom eagerly required the public format files for their
DSL registration side effects, and those files required model.rb
back to ensure the base was loaded.

Each format now splits into a public entrypoint and an internal
implementation:

- lib/lutaml/<format>.rb (public): requires model, then the internal
  file. Standalone require of a single format keeps working.
- lib/lutaml/<format>/format.rb (internal): pure namespace
  definitions plus the registration side effects, and never requires
  model. model.rb's bottom and its Lutaml-level autoloads point at
  these internal files, so the base load never re-enters a public
  entrypoint.
- Sibling references between formats (json -> key_value, toml ->
  key_value) also point at the internal files.

No new lib/lutaml.rb is introduced: the Lutaml root namespace is
shared by other lutaml-* gems and must not be owned by this gem.

The Opal boot manifest is regenerated for the new autoload paths.

Verified by spec/lutaml/require_load_order_spec.rb: ruby -w reports
zero circular-require warnings for require 'lutaml/model' and for
every standalone format require. Full suite 5344 examples, 0
failures; rubocop clean.
