+++ lib/compat/opal/lutaml_model_boot.rb
@@ -21,22 +21,19 @@
 
 require "lutaml/model"
 require "lutaml/xml"
-require "lutaml/hash_format"
-require "lutaml/json"
-require "lutaml/jsonl"
 require "lutaml/jsonld"
-require "lutaml/key_value"
-require "lutaml/toml"
 require "lutaml/turtle"
-require "lutaml/yaml"
 require "lutaml/yamlld"
-require "lutaml/yamls"
 require "lutaml/hash_format/adapter"
+require "lutaml/hash_format/format"
+require "lutaml/json/format"
 require "lutaml/jsonl/adapter"
+require "lutaml/jsonl/format"
 require "lutaml/jsonld/adapter"
 require "lutaml/key_value/adapter"
 require "lutaml/key_value/data_model"
 require "lutaml/key_value/document"
+require "lutaml/key_value/format"
 require "lutaml/key_value/mapping"
 require "lutaml/key_value/mapping_rule"
 require "lutaml/key_value/transform"
@@ -121,6 +118,7 @@ require "lutaml/rdf/namespace_set"
 require "lutaml/rdf/namespaces"
 require "lutaml/rdf/term_definition"
 require "lutaml/rdf/transform"
+require "lutaml/toml/format"
 require "lutaml/turtle/adapter"
 require "lutaml/turtle/mapping"
 require "lutaml/turtle/transform"
@@ -143,6 +141,7 @@ require "lutaml/xml/document"
 require "lutaml/xml/element"
 require "lutaml/xml/element_prefix_resolver"
 require "lutaml/xml/encoding_normalizer"
+require "lutaml/xml/format"
 require "lutaml/xml/format_chooser"
 require "lutaml/xml/hoisting_algorithm"
 require "lutaml/xml/listener"
@@ -185,9 +184,11 @@ require "lutaml/xml/xml_attribute"
 require "lutaml/xml/xml_element"
 require "lutaml/xml/xsd_validator"
 require "lutaml/yaml/adapter"
+require "lutaml/yaml/format"
 require "lutaml/yaml/schema"
 require "lutaml/yamlld/adapter"
 require "lutaml/yamls/adapter"
+require "lutaml/yamls/format"
 require "lutaml/hash_format/adapter/document"
 require "lutaml/hash_format/adapter/mapping"
 require "lutaml/hash_format/adapter/mapping_rule"
